separates chatController tests into integration and unit
